refactor: Refactor scripts code
