+++ packages/sdk/scripts/assets/updateAssets.ts
@@ -1,0 +1,16 @@
+// Script that updates asset map for compatible nodes
+
+import { handleDataFetching } from '../scriptUtils'
+import { fetchAllNodesAssets } from './fetchAssets'
+import { fillDecimalsBifrost } from './fillInDecimalsBifrost'
+
+const JSON_FILE_PATH = './src/maps/assets.json'
+
+;(async () => {
+  await handleDataFetching(
+    JSON_FILE_PATH,
+    fetchAllNodesAssets,
+    'Successfuly updated assets map.',
+    fillDecimalsBifrost
+  )
+})()
